+++ codegen/src/function.rs
@@ -744,7 +744,7 @@ impl ExportedFn {
                                 is_string = true;
                                 is_ref = true;
                                 quote_spanned!(arg_type.span() =>
-                                    mem::take(args[#i]).take_immutable_string().unwrap()
+                                    mem::take(args[#i]).as_immutable_string().unwrap()
                                 )
                             }
                             _ => panic!("internal error: why wasn't this found earlier!?"),
@@ -753,7 +753,7 @@ impl ExportedFn {
                             is_string = true;
                             is_ref = false;
                             quote_spanned!(arg_type.span() =>
-                                mem::take(args[#i]).take_string().unwrap()
+                                mem::take(args[#i]).as_string().unwrap()
                             )
                         }
                         _ => {